Upped warning level to 4 & squashed warnings. Disabled warnings on 3rd party libs as its unactionable noise

+++ Matchstick/include/MGDF/MGDFInputManager.hpp
@@ -103,7 +103,8 @@ IGamepad : public IUnknown {
   controller, valid values are 0-65535 \param left the intensity of the left
   vibrator motor \param right the intensity of the right vibrator motor
   */
-  virtual void STDMETHODCALLTYPE SetVibrationSpeed(INT32 left, INT32 right) = 0;
+  virtual void STDMETHODCALLTYPE SetVibrationSpeed(UINT16 left,
+                                                   UINT16 right) = 0;
 };
 
 /**
